api/requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
django-admin-interface==0.29.4
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-currentuser==0.8.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-redis==5.4.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.9
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.23.1
sqlparse==0.5.3
stripe==11.6.0
text-unidecode==1.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
